Add only_if_torch_available decorator and test_mlflow_hook
